delivery/requirements.txt
django
djangorestframework
uritemplate
pyyaml
pytils
Pillow
django-phone-field
requests
gunicorn 
psycopg2-binary